+++ src/Tracks/TextToSpeech.php
@@ -2,20 +2,19 @@
 
 namespace duncan3dc\Sonos\Tracks;
 
-use duncan3dc\DomParser\XmlWriter;
-use duncan3dc\Helpers\File;
 use duncan3dc\Sonos\Directory;
+use duncan3dc\Sonos\Helper;
 use duncan3dc\Speaker\Providers\GoogleProvider;
 use duncan3dc\Speaker\Providers\ProviderInterface;
 use duncan3dc\Speaker\TextToSpeech as TextToSpeechHandler;
 
 /**
- * Convert a string of a text to a spoken word mp3.
+ * Convert a string of a text to spoken word audio.
  */
 class TextToSpeech implements UriInterface
 {
     /**
-     * @var Directory $directory The directory to store the mp3 in.
+     * @var Directory $directory The directory to store the audio file in.
      */
     protected $directory;
 
@@ -23,11 +22,6 @@ class TextToSpeech implements UriInterface
      * @var string $text The text to convert.
      */
     protected $text;
-
-    /**
-     * @var string $filename The filename of the of the track.
-     */
-    protected $filename;
 
     /**
      * @var Provider $provider The text to speech provider.
@@ -38,20 +32,12 @@ class TextToSpeech implements UriInterface
      * Create a TextToSpeech object.
      *
      * @param string $text The text to convert
-     * @param Directory $directory The directory to store the mp3 in.
+     * @param Directory $directory The directory to store the audio file in
      */
     public function __construct($text, Directory $directory, ProviderInterface $provider = null)
     {
-        if (!is_dir($directory->getFilesystemPath())) {
-            throw new \InvalidArgumentException("Invalid directory: " . $directory->getFilesystemPath());
-        }
-        if (strlen($text) > 100) {
-            throw new \InvalidArgumentException("Only messages under 100 characters are supported");
-        }
-
         $this->directory = $directory;
         $this->text = $text;
-        $this->filename = md5($this->text) . ".mp3";
 
         if ($provider !== null) {
             $this->setProvider($provider);
@@ -95,22 +81,23 @@ class TextToSpeech implements UriInterface
     /**
      * Get the URI for this message.
      *
-     * If it doesn't already exist on the filesystem then the google api will be called.
+     * If it doesn't already exist on the filesystem then the text-to-speech handler will be called.
      *
      * @return string
      */
     public function getUri()
     {
-        $path = $this->directory->getFilesystemPath() . "/{$this->filename}";
+        $provider = $this->getProvider();
+        $tts = new TextToSpeechHandler($this->text, $provider);
 
-        if (!file_exists($path)) {
-            $provider = $this->getProvider();
-            $tts = new TextToSpeechHandler($this->text, $provider);
-            $mp3 = $tts->getAudioData();
-            File::putContents($path, $mp3);
+        $filename = $tts->generateFilename();
+
+        if (!$this->directory->has($filename)) {
+            $data = $tts->getAudioData();
+            $this->directory->write($filename, $data);
         }
 
-        return "x-file-cifs://" . $this->directory->getSharePath() . "/{$this->filename}";
+        return "x-file-cifs://" . $this->directory->getSharePath() . "/{$filename}";
     }
 
 
@@ -121,33 +108,13 @@ class TextToSpeech implements UriInterface
      */
     public function getMetaData()
     {
-        $xml = XmlWriter::createXml([
-            "DIDL-Lite" =>  [
-                "_attributes"   =>  [
-                    "xmlns:dc"      =>  "http://purl.org/dc/elements/1.1/",
-                    "xmlns:upnp"    =>  "urn:schemas-upnp-org:metadata-1-0/upnp/",
-                    "xmlns:r"       =>  "urn:schemas-rinconnetworks-com:metadata-1-0/",
-                    "xmlns"         =>  "urn:schemas-upnp-org:metadata-1-0/DIDL-Lite/",
-                ],
-                "item"  =>  [
-                    "_attributes"   =>  [
-                        "id"            =>  "-1",
-                        "parentID"      =>  "-1",
-                        "restricted"    =>  "true",
-                    ],
-                    "res"               =>  $this->getUri(),
-                    "upnp:albumArtURI"  =>  "",
-                    "dc:title"          =>  $this->text,
-                    "upnp:class"        =>  "object.item.audioItem.musicTrack",
-                    "dc:creator"        =>  "Google",
-                    "upnp:album"        =>  "Text To Speech",
-                ],
-            ]
+        return Helper::createMetaDataXml("-1", "-1", [
+            "res"               =>  $this->getUri(),
+            "upnp:albumArtURI"  =>  "",
+            "dc:title"          =>  $this->text,
+            "upnp:class"        =>  "object.item.audioItem.musicTrack",
+            "dc:creator"        =>  "Google",
+            "upnp:album"        =>  "Text To Speech",
         ]);
-
-        # Get rid of the xml header as only the DIDL-Lite element is required
-        $meta = explode("\n", $xml)[1];
-
-        return $meta;
     }
 }